requirements.txt
flask==3.1.1
python-dotenv==1.1.1
requests==2.32.4
transformers==4.53.2
torch==2.2.2
